LayoutTests/fast/dom/body-margin-dynamic-iframe-propagation.html

<!doctype html>
<meta charset=utf-8>
<title>Dynamic iframe margin attribute propagation to body</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
// Tests that dynamically changing marginwidth/marginheight on an iframe
// correctly propagates to the child document's body element, and respects
// precedence over leftmargin/topmargin on the body.
// HTMLFrameElementBase::attributeChanged handles this propagation.

function createIframe(bodySrc) {
    return new Promise(resolve => {
        const iframe = document.createElement("iframe");
        iframe.src = "data:text/html,<!doctype html>" + encodeURIComponent(bodySrc);
        iframe.onload = () => resolve(iframe);
        document.body.appendChild(iframe);
    });
}

function bodyStyle(iframe) {
    return iframe.contentWindow.getComputedStyle(iframe.contentDocument.body);
}

promise_test(async t => {
    const iframe = await createIframe("<body>");
    t.add_cleanup(() => iframe.remove());
    iframe.setAttribute("marginwidth", "30");
    assert_equals(bodyStyle(iframe).marginLeft, "30px", "marginwidth propagated to margin-left");
    assert_equals(bodyStyle(iframe).marginRight, "30px", "marginwidth propagated to margin-right");
}, "Dynamically setting iframe marginwidth propagates to child body");

promise_test(async t => {
    const iframe = await createIframe("<body>");
    t.add_cleanup(() => iframe.remove());
    iframe.setAttribute("marginheight", "30");
    assert_equals(bodyStyle(iframe).marginTop, "30px", "marginheight propagated to margin-top");
    assert_equals(bodyStyle(iframe).marginBottom, "30px", "marginheight propagated to margin-bottom");
}, "Dynamically setting iframe marginheight propagates to child body");

promise_test(async t => {
    const iframe = await createIframe("<body>");
    t.add_cleanup(() => iframe.remove());
    iframe.setAttribute("marginwidth", "30");
    assert_equals(bodyStyle(iframe).marginLeft, "30px", "initially 30px");
    iframe.setAttribute("marginwidth", "50");
    assert_equals(bodyStyle(iframe).marginLeft, "50px", "updated to 50px");
    iframe.removeAttribute("marginwidth");
    assert_not_equals(bodyStyle(iframe).marginLeft, "50px", "removed marginwidth");
}, "Updating and removing iframe marginwidth updates child body");

promise_test(async t => {
    const iframe = await createIframe("<body leftmargin=10>");
    t.add_cleanup(() => iframe.remove());
    // Body has leftmargin=10, no marginwidth yet on the body.
    assert_equals(bodyStyle(iframe).marginLeft, "10px", "leftmargin=10 initially");
    // Dynamically set marginwidth on the iframe — HTMLFrameElementBase::attributeChanged
    // will set marginwidth on the body. marginwidth takes precedence over leftmargin.
    iframe.setAttribute("marginwidth", "30");
    assert_equals(bodyStyle(iframe).marginLeft, "30px", "iframe marginwidth takes precedence over body leftmargin");
    iframe.removeAttribute("marginwidth");
    assert_equals(bodyStyle(iframe).marginLeft, "10px", "body leftmargin restored after iframe marginwidth removed");
}, "Dynamic iframe marginwidth takes precedence over body leftmargin");

promise_test(async t => {
    const iframe = await createIframe("<body topmargin=10>");
    t.add_cleanup(() => iframe.remove());
    assert_equals(bodyStyle(iframe).marginTop, "10px", "topmargin=10 initially");
    iframe.setAttribute("marginheight", "30");
    assert_equals(bodyStyle(iframe).marginTop, "30px", "iframe marginheight takes precedence over body topmargin");
    iframe.removeAttribute("marginheight");
    assert_equals(bodyStyle(iframe).marginTop, "10px", "body topmargin restored after iframe marginheight removed");
}, "Dynamic iframe marginheight takes precedence over body topmargin");
</script>
</body>
